strnstr

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: mriant <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/30 14:55:36 by mriant            #+#    #+#             */
-/*   Updated: 2021/11/30 15:10:02 by mriant           ###   ########.fr       */
+/*   Updated: 2021/12/02 09:33:56 by mriant           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	s_len;
+	int	s_len;
 
 	s_len = ft_strlen(s);
 	while (s_len >= 0)
